+++ apps/site/tests/e2e/api.test.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "../shared/runtime.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 test("/api/me should be healthy", async ({ request }) => {
   // This test helps check DB connection. If the endpoint returns 500,
