+++ apps/site/tests/integration/accountpage.test.ts
@@ -1,6 +1,6 @@
 import { getBlocksData, resetSite } from "../shared/fixtures.js";
 import { login } from "../shared/nav.js";
-import { expect, test } from "../shared/runtime.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 test.beforeEach(async () => {
   await resetSite();
