+++ apps/site/tests/integration/block-publish-npm.test.ts
@@ -1,8 +1,8 @@
 import { publishBlock } from "../shared/blocks.js";
 import { resetSite } from "../shared/fixtures.js";
 import { login } from "../shared/nav.js";
-import type { Page } from "../shared/runtime.js";
-import { expect, test } from "../shared/runtime.js";
+import type { Page } from "../shared/wrapped-playwright.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 const fillBlockDetails = async (
   page: Page,
