+++ apps/site/tests/integration/blockpage.test.ts
@@ -1,5 +1,5 @@
 import { getBlocksData } from "../shared/fixtures.js";
-import { expect, test } from "../shared/runtime.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 const blocksData = await getBlocksData();
