+++ apps/site/tests/integration/page-footer.test.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "../shared/runtime.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 test("page footer navigation works", async ({ page, browserName }) => {
   test.skip(
