+++ apps/site/tests/integration/schema-creation.test.ts
@@ -1,7 +1,7 @@
 import { resetSite } from "../shared/fixtures.js";
 import { login } from "../shared/nav.js";
-import { expect, test } from "../shared/runtime.js";
 import { createSchema } from "../shared/schemas.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 test("user should be able to create schema", async ({ page }) => {
   await resetSite();
