+++ apps/site/tests/integration/schemapage.test.ts
@@ -1,7 +1,7 @@
 import { resetSite } from "../shared/fixtures.js";
 import { login } from "../shared/nav.js";
-import { expect, test } from "../shared/runtime.js";
 import { createSchema } from "../shared/schemas.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 test.beforeEach(async () => {
   await resetSite();
