+++ apps/site/tests/integration/search-modal.test.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "../shared/runtime.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 const searchModalSelector = '[data-testid="bp-search-modal"]';
