+++ apps/site/tests/integration/sign-up-flow.ts
@@ -1,7 +1,7 @@
 import { readValueFromRecentDummyEmail } from "../shared/dummy-emails.js";
 import { resetSite } from "../shared/fixtures.js";
 import { login, openMobileNav } from "../shared/nav.js";
-import { expect, test } from "../shared/runtime.js";
+import { expect, test } from "../shared/wrapped-playwright.js";
 
 test("sign up flow works", async ({ browserName, isMobile, page }) => {
   await resetSite();
