+++ apps/site/tests/shared/nav.ts
@@ -1,6 +1,6 @@
 import { readValueFromRecentDummyEmail } from "./dummy-emails.js";
-import type { Locator, Page } from "./runtime.js";
-import { expect } from "./runtime.js";
+import type { Locator, Page } from "./wrapped-playwright.js";
+import { expect } from "./wrapped-playwright.js";
 
 export const openMobileNav = async (page: Page) => {
   await page.locator("[data-testid='mobile-nav-trigger']").click();
