+++ apps/site/tests/shared/schemas.ts
@@ -1,6 +1,6 @@
 import type { EntityType } from "../../src/lib/api/model/entity-type.model";
-import type { Page } from "./runtime.js";
-import { expect } from "./runtime.js";
+import type { Page } from "./wrapped-playwright.js";
+import { expect } from "./wrapped-playwright.js";
 
 /**
  * Provides a quick way to create entity types using direct API calls instead of UI
